+++ src/eppo-client.ts
@@ -2,6 +2,8 @@ import { createHash } from 'crypto';
 
 import { IExperimentConfiguration } from './experiment/experiment-configuration';
 import ExperimentConfigurationRequestor from './experiment/experiment-configuration-requestor';
+import { Rule, AttributeValueType } from './rule';
+import { matchesAnyRule } from './rule_evaluator';
 import { getShard, isShardInRange } from './shard';
 import { validateNotBlank } from './validation';
 
@@ -13,50 +15,58 @@ export interface IEppoClient {
   /**
    * Maps a subject to a variation for a given experiment.
    *
-   * @param subject an entity ID, e.g. userId
+   * @param subjectKey an identifier of the experiment subject, for example a user ID.
    * @param experimentKey experiment identifier
+   * @param subjectAttributes optional attributes associated with the subject, for example name and email.
+   * The subject attributes are used for evaluating any targeting rules tied to the experiment.
    * @returns a variation value if the subject is part of the experiment sample, otherwise null
    * @public
    */
-  getAssignment(subject: string, experimentKey: string): string;
-
-  /**
-   * Returns a Promise that resolves once the client polling process has started.
-   * @public
-   */
-  waitForInitialization: () => Promise<void>;
+  getAssignment(
+    subjectKey: string,
+    experimentKey: string,
+    subjectAttributes?: Record<string, AttributeValueType>,
+  ): string;
 }
 
 export default class EppoClient implements IEppoClient {
-  constructor(
-    public waitForInitialization: () => Promise<void>,
-    private configurationRequestor: ExperimentConfigurationRequestor,
-  ) {}
+  constructor(private configurationRequestor: ExperimentConfigurationRequestor) {}
 
-  getAssignment(subject: string, experimentKey: string): string {
-    validateNotBlank(subject, 'Invalid argument: subject cannot be blank');
+  getAssignment(subjectKey: string, experimentKey: string, subjectAttributes = {}): string {
+    validateNotBlank(subjectKey, 'Invalid argument: subjectKey cannot be blank');
     validateNotBlank(experimentKey, 'Invalid argument: experimentKey cannot be blank');
     const experimentConfig = this.configurationRequestor.getConfiguration(experimentKey);
     if (
       !experimentConfig?.enabled ||
-      !this.isInExperimentSample(subject, experimentKey, experimentConfig)
+      !this.subjectAttributesSatisfyRules(subjectAttributes, experimentConfig.rules) ||
+      !this.isInExperimentSample(subjectKey, experimentKey, experimentConfig)
     ) {
       return null;
     }
-    const override = this.getSubjectVariationOverride(subject, experimentConfig);
+    const override = this.getSubjectVariationOverride(subjectKey, experimentConfig);
     if (override) {
       return override;
     }
     const { variations, subjectShards } = experimentConfig;
-    const shard = getShard(`assignment-${subject}-${experimentKey}`, subjectShards);
+    const shard = getShard(`assignment-${subjectKey}-${experimentKey}`, subjectShards);
     return variations.find((variation) => isShardInRange(shard, variation.shardRange)).name;
   }
 
+  private subjectAttributesSatisfyRules(
+    subjectAttributes?: Record<string, AttributeValueType>,
+    rules?: Rule[],
+  ) {
+    if (!rules || rules.length === 0) {
+      return true;
+    }
+    return matchesAnyRule(subjectAttributes || {}, rules);
+  }
+
   private getSubjectVariationOverride(
-    subject: string,
+    subjectKey: string,
     experimentConfig: IExperimentConfiguration,
   ): string {
-    const subjectHash = createHash('md5').update(subject).digest('hex');
+    const subjectHash = createHash('md5').update(subjectKey).digest('hex');
     return experimentConfig.overrides[subjectHash];
   }
 
@@ -66,12 +76,12 @@ export default class EppoClient implements IEppoClient {
    * Given a hash function output (bucket), check whether the bucket is between 0 and exposure_percent * total_buckets.
    */
   private isInExperimentSample(
-    subject: string,
+    subjectKey: string,
     experimentKey: string,
     experimentConfig: IExperimentConfiguration,
   ): boolean {
     const { percentExposure, subjectShards } = experimentConfig;
-    const shard = getShard(`exposure-${subject}-${experimentKey}`, subjectShards);
+    const shard = getShard(`exposure-${subjectKey}-${experimentKey}`, subjectShards);
     return shard <= percentExposure * subjectShards;
   }
 }