+++ CHANGELOG.md
@@ -11,6 +11,19 @@
 * Describe deprecated APIs in this version
 -->
 
+## [1.0.0] - 2022-06-03
+
+#### New Features:
+* Subject attributes: an optional `subjectAttributes` param is added to the `getAssignment` function. The subject attributes may contains custom metadata about the subject. These attributes are used for evaluating any targeting rules defined on the experiment.
+```
+client.getAssignment("<SUBJECT_KEY">, "<EXPERIMENT_KEY>", { "email": "user@example.com" });
+```
+#### Breaking Changes:
+* The `init` method is changed from sync to async:
+```
+await init({ apiKey: "<API_KEY>" });
+```
+
 ## [0.4.0] - 2022-05-12
 
 #### New Features: